Renaming 'attrib' to 'attribute'
