+++ plugins/services/src/js/filters/ServiceHasVolumesAttributeFilter.js
@@ -1,17 +1,16 @@
 import DSLFilterTypes from '../../../../../src/js/constants/DSLFilterTypes';
 import DSLFilter from '../../../../../src/js/structs/DSLFilter';
-import Framework from '../structs/Framework';
 
-const LABEL = 'is';
-const LABEL_TEXT = 'universe';
+const LABEL = 'has';
+const LABEL_TEXT = 'volumes';
 
 /**
- * This filter handles the `is:status` for services using `getServiceStatus`
+ * This filter handles the `has:volumes` for filtering services with volumes
  */
-class ServiceIsUniverseAttribFilter extends DSLFilter {
+class ServiceIsUniverseAttributeFilter extends DSLFilter {
 
   /**
-   * Handle all `is:universe` attrib filters that we can handle.
+   * Handle all `has:volumes` attrib filters that we can handle.
    *
    * @override
    */
@@ -22,16 +21,17 @@ class ServiceIsUniverseAttribFilter extends DSLFilter {
   }
 
   /**
-   * Keep only services whose instance is of type `Framework`
+   * Keep only services that contain more than 1 volume
    *
    * @override
    */
   filterApply(resultset) {
     return resultset.filterItems((service) => {
-      return service instanceof Framework;
+      let volumes = service.getVolumes();
+      return (volumes.list && volumes.list.length > 0);
     });
   }
 
 }
 
-module.exports = ServiceIsUniverseAttribFilter;
+module.exports = ServiceIsUniverseAttributeFilter;
